added dead-end & branch counter, organized code more by splitting up customizemaze.js into itself & playmaze.js.
