+++ index.html
@@ -194,7 +194,7 @@
             Download
           </button>
         </div>
-        
+        <div id="deadEndCount"></div>
       </div>
       <div id="time" class="container mt-5 time mx-auto text-center"></div>
       <div class="container my-5 d-flex justify-content-center">
@@ -225,6 +225,7 @@
     <script type="module" src="src/grid.js"></script>
     <script type="module" src="src/maze.js"></script>
     <script type="module" src="src/customizemaze.js"></script>
+    <script type="module" src="src/playmaze.js"></script>
   </script>
 </body>
 </html>
